6-size.c

+++ 0x00-hello_world/6-size.c
@@ -15,7 +15,7 @@ int main(void)
 	char c;
 
 	printf("Size of a char: %ld byte(s) \n", sizeof(c));
-	printf("Size of an int: %lu byte(s) \n", (unsigned long) sizeof(i));
+	printf("Size of an int: %ld byte(s) \n", sizeof(i));
         printf("Size of a long int: %ld byte(s) \n", sizeof(lint));
         printf("Size of a long long int: %ld byte(s) \n", sizeof(llint));
         printf("Size of a float: %ld byte(s) \n", sizeof(f));
